requirements.txt
python-dotenv
pydub
rpi.gpio
